backend/requirements.txt

flask
flask-cors
numpy
requests
pillow
Flask
flask-cors
numpy
Pillow

torch
torchvision
transformers
facenet-pytorch

opencv-python
mediapipe
timm

librosa

tldextract
beautifulsoup4

librosa
requests
tldextract
beautifulsoup4
opencv-python
mediapipe

